pyproject.toml
[tool.poetry]
name = "wishlistr-api"
version = "0.1.0"
description = ""
authors = ["Lasse Nielsen <lasse@xerrion.dk>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = "*"
argon2-cffi = "*"
python-jose = {extras = ["cryptography"], version = "*"}
passlib = "*"
sqlalchemy = "*"
pydantic = "*"
python-multipart = "*"
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "*"